+++ semaphoreci.sh
@@ -20,6 +20,7 @@ test() {
 
 test_docs() {
 	$STACK --no-terminal ghc -- -Wall -fno-code docs/tutorial/tutorial.lhs
+        $STACK --no-terminal haddock --no-haddock-deps
 }
 
 setup_lint() {